Switch Group: line to newer tag proposal
"Classic groups" (with the slash separator) put a package in one
group, which has been regarded as a limitation. A tag-based approach
is instead able to do a multi-group placement similar to XDG
categories. This commit moves cpanspec to emit this new syntax,
starting off with the two tags "development" and "perl" (similar to,
but not the same as the classic group Development/Languages/Perl),
dropping the unspecific "languages" tag ("perl" is a better fit).